A2Q1 changed int to long

+++ A1/A1/src/A2Q1/SortedIntegerArray.java
@@ -29,7 +29,7 @@ public class SortedIntegerArray {
     }
     
     private boolean kPairSumInterval(Integer k, int i, int j) {
-    	int tempSum = sortedIntegerArray[i] + sortedIntegerArray[j];
+    	long tempSum = sortedIntegerArray[i] + sortedIntegerArray[j];
     	int newI = i+1;
     	int newJ = j-1;
     	if (tempSum == k) {
